Setup module for test helpers
